requirements.txt
aiofiles
fastapi
sqlalchemy
dotenv
asyncpg
python-multipart
requests
uvicorn
httpx
docxcomposeпше